# api/serverpb/rpc.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "api/serverpb/rpc.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "Resources"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/healthz": {
      "get": {
        "operationId": "Resources_Healthz",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/serverpbHealthzResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "Resources"
        ]
      }
    },
    "/leader": {
      "get": {
        "operationId": "Resources_GetLeader",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/serverpbGetLeaderResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "Resources"
        ]
      }
    },
    "/v1/resources/diff": {
      "post": {
        "operationId": "Resources_Diff",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/serverpbDiffResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/serverpbDiffRequest"
            }
          }
        ],
        "tags": [
          "Resources"
        ]
      }
    },
    "/v1/resources/list": {
      "post": {
        "operationId": "Resources_List",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/serverpbListResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/serverpbListRequest"
            }
          }
        ],
        "tags": [
          "Resources"
        ]
      }
    },
    "/v1/resources/sync": {
      "post": {
        "operationId": "Resources_Sync",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/serverpbSyncResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/serverpbSyncRequest"
            }
          }
        ],
        "tags": [
          "Resources"
        ]
      }
    }
  },
  "definitions": {
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "protobufNullValue": {
      "type": "string",
      "enum": [
        "NULL_VALUE"
      ],
      "default": "NULL_VALUE",
      "description": "`NullValue` is a singleton enumeration to represent the null value for the\n`Value` type union.\n\n The JSON representation for `NullValue` is JSON `null`.\n\n - NULL_VALUE: Null value."
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "serverpbDiffRequest": {
      "type": "object",
      "properties": {
        "gateway": {
          "type": "string",
          "description": "网关名",
          "title": "gateway"
        },
        "stage": {
          "type": "string",
          "description": "环境名",
          "title": "stage"
        },
        "resource": {
          "$ref": "#/definitions/serverpbResourceIdentity"
        },
        "all": {
          "type": "boolean",
          "description": "所有网关环境",
          "title": "all"
        }
      }
    },
    "serverpbDiffResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int64",
          "description": "返回错误码",
          "title": "code"
        },
        "message": {
          "type": "string",
          "description": "返回错误信息",
          "title": "message"
        },
        "data": {
          "type": "object",
          "additionalProperties": {
            "$ref": "#/definitions/serverpbStageScopedApisixResources"
          },
          "description": "diff 内容，以gateway/stage为key",
          "title": "message"
        }
      }
    },
    "serverpbGetLeaderResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int64",
          "description": "返回错误码",
          "title": "code"
        },
        "message": {
          "type": "string",
          "description": "返回错误信息",
          "title": "message"
        },
        "data": {
          "type": "string",
          "description": "leader 实例名",
          "title": "data"
        }
      }
    },
    "serverpbHealthzResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int64",
          "description": "返回错误码",
          "title": "code"
        },
        "message": {
          "type": "string",
          "description": "返回错误信息",
          "title": "message"
        }
      }
    },
    "serverpbListRequest": {
      "type": "object",
      "properties": {
        "gateway": {
          "type": "string",
          "description": "网关名",
          "title": "gateway"
        },
        "stage": {
          "type": "string",
          "description": "环境名",
          "title": "stage"
        },
        "resource": {
          "$ref": "#/definitions/serverpbResourceIdentity"
        },
        "all": {
          "type": "boolean",
          "description": "所有网关环境",
          "title": "all"
        }
      }
    },
    "serverpbListResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int64",
          "description": "返回错误码",
          "title": "code"
        },
        "message": {
          "type": "string",
          "description": "返回错误信息",
          "title": "message"
        },
        "data": {
          "type": "object",
          "additionalProperties": {
            "$ref": "#/definitions/serverpbStageScopedApisixResources"
          },
          "description": "apisix 资源内容，以gateway/stage为key",
          "title": "message"
        }
      }
    },
    "serverpbResourceIdentity": {
      "type": "object",
      "properties": {
        "resourceId": {
          "type": "string",
          "format": "int64"
        },
        "resourceName": {
          "type": "string"
        }
      }
    },
    "serverpbStageScopedApisixResources": {
      "type": "object",
      "properties": {
        "routes": {
          "type": "object"
        },
        "services": {
          "type": "object"
        },
        "pluginMetadata": {
          "type": "object"
        },
        "ssl": {
          "type": "object"
        }
      }
    },
    "serverpbSyncRequest": {
      "type": "object",
      "properties": {
        "gateway": {
          "type": "string",
          "description": "网关名",
          "title": "gateway"
        },
        "stage": {
          "type": "string",
          "description": "环境名",
          "title": "stage"
        },
        "all": {
          "type": "boolean",
          "description": "所有网关环境",
          "title": "all"
        }
      }
    },
    "serverpbSyncResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int64",
          "description": "返回错误码",
          "title": "code"
        },
        "message": {
          "type": "string",
          "description": "返回错误信息",
          "title": "message"
        }
      }
    }
  }
}